requirements.txt
ldap3
python-dotenv
Flask
Flask-Cors
Flask-Mail
Flask-Limiter
pytest
pycryptodome
requests